Refactor type names and files
