+++ src/dealer/HedgingStrategyFactory.ts
@@ -1,22 +1,19 @@
-import { IHedgingStrategy, HedgingStrategies } from "./IHedgingStrategy"
+import { HedgingStrategy, HedgingStrategies } from "./HedgingStrategyTypes"
 import { FtxPerpetualSwapStrategy } from "./FtxPerpetualSwapStrategy"
 import { OkexPerpetualSwapStrategy } from "./OkexPerpetualSwapStrategy"
 
 export function createHedgingStrategy(
   strategy: HedgingStrategies,
   logger,
-): IHedgingStrategy {
+): HedgingStrategy {
   switch (strategy) {
     case HedgingStrategies.FtxPerpetualSwap:
       return new FtxPerpetualSwapStrategy(logger)
-      break
 
     case HedgingStrategies.OkexPerpetualSwap:
       return new OkexPerpetualSwapStrategy(logger)
-      break
 
     default:
       throw new Error("Not implemented!")
-      break
   }
 }
